requirements.txt
ttkbootstrap
transformers
torch
keybert
langdetect
python-docx
PyPDF2